[refactor] OuputFormula 분리
- OuutputFormula = OutputFormula + OutputFormulaBox 로 분리
- OuputFormulaBox는 MenuBar에 재사용
